pyproject.toml
[project]
name = "gilat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "llama-cloud-services>=0.6.12",
    "llama-index>=0.12.30",
    "llama-index-core>=0.12.30",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-openai>=0.3.33",
    "llama-index-readers-file>=0.4.7",
    "nest-asyncio>=1.6.0",
    "python-decouple>=3.8",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
